requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.110.1
h11==0.14.0
idna==3.6
mongoengine==0.28.2
openpyxl==3.1.2
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
uvicorn==0.29.0
